toolkit/mozapps/downloads/tests/chrome/test_privatebrowsing_title.xul

<?xml version="1.0"?>
<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/.  -->
<!--
 * Test bug 464800 to make sure that the download manager window title is
 * properly cleared when transitioning into the private browsing mode.
-->

<window title="Private Browsing transition title change test"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="test();">

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"/>
  <script type="application/javascript"
          src="utils.js"/>

  <script type="application/javascript">
  <![CDATA[

function test()
{
  // The private browsing service might not be available
  let pbs = null;
  try {
    pbs = Cc["@mozilla.org/privatebrowsing;1"].getService(Ci.nsIPrivateBrowsingService);
  } catch (e) {
  }
  if (!pbs) {
    todo(false, "skip test - no private browsing supported");
    return;
  }

  var dmui = getDMUI();
  if (!dmui) {
    todo(false, "skip test for toolkit download manager UI");
    return;
  }

  // We should prevent the private browsing service from changing the current
  // session on private mode transitions, in order to play nice with the
  // Mochitest test harness.
  let prefBranch = Cc["@mozilla.org/preferences-service;1"].
                   getService(Ci.nsIPrefBranch);
  prefBranch.setBoolPref("browser.privatebrowsing.keep_current_session", true);

  // Start with an empty downloads database
  Cc["@mozilla.org/download-manager;1"].
  getService(Ci.nsIDownloadManager).
  DBConnection.
  executeSimpleSQL("DELETE FROM moz_downloads");

  // Close the UI if necessary
  let wm = Cc["@mozilla.org/appshell/window-mediator;1"].
           getService(Ci.nsIWindowMediator);
  let win = wm.getMostRecentWindow("Download:Manager");
  if (win) win.close();

  let obs = Cc["@mozilla.org/observer-service;1"].
            getService(Ci.nsIObserverService);
  const DLMGR_UI_DONE = "download-manager-ui-done";

  let testObs = {
    observe: function(aSubject, aTopic, aData) {
      if (aTopic != DLMGR_UI_DONE)
        return;
      obs.removeObserver(testObs, DLMGR_UI_DONE);

      let win = aSubject.QueryInterface(Ci.nsIDOMWindow);
      win.focus();
      let defaultTitle = win.document.documentElement.getAttribute("statictitle");
      is(win.document.title, defaultTitle,
        "The downloads window title is correct outside of the private browsing mode");
      win.document.title = "some title";

      // Enter the private browsing mode
      pbs.privateBrowsingEnabled = true;

      is(win.document.title, defaultTitle,
        "The downloads window title is correct inside the private browsing mode");
      win.document.title = "another title";

      // Leave the private browsing mode
      pbs.privateBrowsingEnabled = false;

      is(win.document.title, defaultTitle,
        "The downloads window title is correct after leaving the private browsing mode");

      // We're done!
      win.close();
      prefBranch.clearUserPref("browser.privatebrowsing.keep_current_session");
      SimpleTest.finish();
    }
  };
  obs.addObserver(testObs, DLMGR_UI_DONE, false);

  // Show the Download Manager UI
  dmui.show();

  SimpleTest.waitForExplicitFinish();
}

  ]]>
  </script>

  <body xmlns="http://www.w3.org/1999/xhtml">
    <p id="display"></p>
    <div id="content" style="display:none;"></div>
    <pre id="test"></pre>
  </body>
</window>
